requirements.txt
# ==============================================================================
# CORE DJANGO - Essentiels pour GreenCart MVP
# ==============================================================================
Django==5.2.4
python-decouple==3.8
dj-database-url==2.1.0

# ==============================================================================
# API & REST FRAMEWORK - Pour l'API GreenCart
# ==============================================================================
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0

# ==============================================================================
# API DOCUMENTATION - Pour Swagger/OpenAPI
# ==============================================================================
drf-spectacular==0.27.2

# ==============================================================================
# IMAGES & MEDIA - Pour les photos des produits
# ==============================================================================
Pillow==10.4.0

# ==============================================================================
# DATABASE & PRODUCTION - Pour PostgreSQL et déploiement
# ==============================================================================
psycopg2-binary==2.9.10
gunicorn==21.2.0
whitenoise==6.5.0